requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
djangorestframework==3.14.0
idna==3.4
psycopg==3.1.7
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
